% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_lookups.r
\name{cd_to_zip}
\alias{cd_to_zip}
\title{Get all ZIP codes that fall within a given congressional district}
\usage{
cd_to_zip(state_fips_code, congressional_district)
}
\arguments{
\item{state_fips_code}{A two-digit U.S. FIPS code for a state}

\item{congressional_district}{A two digit number specifying a congressional district in a given}
}
\value{
dataframe of all congressional districts found for given ZIP code, including state code
}
\description{
Get all ZIP codes that fall within a given congressional district
}
\examples{
cd_to_zip('34','03')
cd_to_zip('36','05')
}
