% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng_functions.R
\name{get_cd_ng}
\alias{get_cd_ng}
\title{Get congressional districts using an explicit data bundle}
\usage{
get_cd_ng(data, zip_code)
}
\arguments{
\item{data}{A \code{zipcodeR_data_bundle}.}

\item{zip_code}{One five-digit ZIP code.}
}
\value{
A named list with parallel \code{state_fips} and \code{district} vectors.
}
\description{
Get congressional districts using an explicit data bundle
}
