% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_lookups.r
\name{get_con_dist}
\alias{get_con_dist}
\title{Get all congressional districts for a given ZIP code}
\usage{
get_con_dist(zip_code)
}
\arguments{
\item{zip_code}{A U.S. ZIP code}
}
\value{
dataframe of all congressional districts found for given ZIP code, including state code
}
\description{
Get all congressional districts for a given ZIP code
}
\examples{
get_con_dist('08731')
get_con_dist('90210')
}
