% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_lookups.r
\name{get_tracts}
\alias{get_tracts}
\title{Get all Census tracts within a given ZIP code}
\usage{
get_tracts(zip_code)
}
\arguments{
\item{zip_code}{A U.S. ZIP code}
}
\value{
dataframe of Census tracts and data from Census crosswalk file found for given ZIP code
}
\description{
Get all Census tracts within a given ZIP code
}
\examples{
get_tracts('08731')
get_tracts('90210')
}
