% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng_functions.R
\name{get_tracts_ng}
\alias{get_tracts_ng}
\title{Get Census tracts using an explicit data bundle}
\usage{
get_tracts_ng(data, zip_code)
}
\arguments{
\item{data}{A \code{zipcodeR_data_bundle}.}

\item{zip_code}{One five-digit ZIP or ZCTA code.}
}
\value{
A tract crosswalk tibble.
}
\description{
Get Census tracts using an explicit data bundle
}
