% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_lookups.r
\name{is_zcta}
\alias{is_zcta}
\title{Returns true if the given ZIP code is also a ZIP code tabulation area (ZCTA)}
\usage{
is_zcta(zip_code)
}
\arguments{
\item{zip_code}{A 5-digit U.S. ZIP code}
}
\value{
Boolean TRUE or FALSE based upon whether provided ZIP code is a ZCTA by testing whether it exists in the U.S. Census crosswalk data
}
\description{
Returns true if the given ZIP code is also a ZIP code tabulation area (ZCTA)
}
\examples{
is_zcta("90210")
is_zcta("99999")
is_zcta("07762")
}
