% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng_functions.R
\name{is_zcta_ng}
\alias{is_zcta_ng}
\title{Test ZCTA membership using an explicit data bundle}
\usage{
is_zcta_ng(data, zip_code)
}
\arguments{
\item{data}{A \code{zipcodeR_data_bundle}.}

\item{zip_code}{ZIP code(s).}
}
\value{
A logical vector.
}
\description{
Test ZCTA membership using an explicit data bundle
}
