% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_helper_functions.R
\name{normalize_zip}
\alias{normalize_zip}
\title{Normalize ZIP codes}
\usage{
normalize_zip(zipcode)
}
\arguments{
\item{zipcode}{messy ZIP code to be normalized}
}
\value{
Normalized zipcode
}
\description{
Normalize ZIP codes
}
\examples{
normalize_zip(0008731)
}
