% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_lookups.r
\name{search_cd}
\alias{search_cd}
\title{Get all ZIP codes that fall within a given congressional district}
\usage{
search_cd(state_fips_code, congressional_district)
}
\arguments{
\item{state_fips_code}{A two-digit U.S. FIPS code for a state}

\item{congressional_district}{A two digit number specifying a congressional district in a given}
}
\value{
tibble of all congressional districts found for given ZIP code, including state code
}
\description{
Get all ZIP codes that fall within a given congressional district
}
\examples{
search_cd("34", "03")
search_cd("36", "05")
}
