% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng_functions.R
\name{search_cd_ng}
\alias{search_cd_ng}
\title{Search a congressional district using an explicit data bundle}
\usage{
search_cd_ng(data, state_fips_code, congressional_district)
}
\arguments{
\item{data}{A \code{zipcodeR_data_bundle}.}

\item{state_fips_code}{One state FIPS code.}

\item{congressional_district}{One two-digit district code.}
}
\value{
A ZIP-to-district tibble.
}
\description{
Search a congressional district using an explicit data bundle
}
