% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng_functions.R
\name{search_city_ng}
\alias{search_city_ng}
\title{Search a city using an explicit data bundle}
\usage{
search_city_ng(data, city_name, state_abb)
}
\arguments{
\item{data}{A \code{zipcodeR_data_bundle}.}

\item{city_name}{City name.}

\item{state_abb}{Two-letter state abbreviation.}
}
\value{
A tibble from the selected data vintage.
}
\description{
Search a city using an explicit data bundle
}
