% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_lookups.r
\name{search_county}
\alias{search_county}
\title{Search ZIP codes for a county}
\usage{
search_county(county_name, state_abb)
}
\arguments{
\item{county_name}{Name of a county within a U.S. state}

\item{state_abb}{Two-digit code for a U.S. state}
}
\value{
dataframe of all ZIP codes found for given county
}
\description{
Search ZIP codes for a county
}
\examples{
middlesex <- search_county('Middlesex','NJ')
alameda <- search_county('alameda','CA')
}
