% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng_functions.R
\name{search_fips_ng}
\alias{search_fips_ng}
\title{Search FIPS codes using an explicit data bundle}
\usage{
search_fips_ng(data, state_fips, county_fips)
}
\arguments{
\item{data}{A \code{zipcodeR_data_bundle}.}

\item{state_fips}{One state FIPS code.}

\item{county_fips}{Optional one- to three-digit county FIPS code.}
}
\value{
A tibble from the selected data vintage.
}
\description{
Search FIPS codes using an explicit data bundle
}
