% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_lookups.r
\name{search_point}
\alias{search_point}
\title{Search for ZIP codes that are within a given radius from a point}
\usage{
search_point(lat, lng, radius = 1)
}
\arguments{
\item{lat}{latitude}

\item{lng}{longitude}

\item{radius}{distance to search in miles, set by default to 1}
}
\value{
ZIP code(s)
}
\description{
Search for ZIP codes that are within a given radius from a point
}
\examples{
search_point(39.9, -74.3, 10)
}
