% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_lookups.r
\name{search_radius}
\alias{search_radius}
\title{Search for ZIP codes that are within a given radius from a point}
\usage{
search_radius(lat, lng, radius = 1)
}
\arguments{
\item{lat}{latitude}

\item{lng}{longitude}

\item{radius}{distance to search in miles, set by default to 1}
}
\value{
a tibble containing the ZIP code(s) within the provided radius and distance from the provided coordinates in miles
}
\description{
Search for ZIP codes that are within a given radius from a point
}
\examples{

\dontrun{search_radius(39.9, -74.3, 10)}
}
