% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng_functions.R
\name{search_radius_ng}
\alias{search_radius_ng}
\title{Search within a radius using an explicit data bundle}
\usage{
search_radius_ng(data, lat, lng, radius = 1)
}
\arguments{
\item{data}{A \code{zipcodeR_data_bundle}.}

\item{lat, lng}{Query coordinates.}

\item{radius}{Radius in miles.}
}
\value{
A tibble of ZIP codes and haversine distances.
}
\description{
Search within a radius using an explicit data bundle
}
