% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_lookups.r
\name{search_state}
\alias{search_state}
\title{Search for ZIP codes located within a given state}
\usage{
search_state(state_abb)
}
\arguments{
\item{state_abb}{Two-digit code representing a U.S. state}
}
\value{
tibble of all ZIP codes for each state code defined in state_abb
}
\description{
Search for ZIP codes located within a given state
}
\examples{
search_state("NJ")
search_state(c("NJ", "NY", "CT"))
}
