% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{zcta_crosswalk}
\alias{zcta_crosswalk}
\title{ZCTA to Census Tract (2020) Crosswalk}
\format{
A data frame with 168212 rows and 3 variables:
\describe{
\item{ZCTA5}{2020 ZIP Code Tabulation Area}
\item{TRACT}{2020 Census Tract Code}
\item{GEOID}{Concatenation of 2020 State, County, and Tract}
}
}
\source{
\url{https://www.census.gov/geographies/reference-files/time-series/geo/relationship-files.html}
}
\usage{
zcta_crosswalk
}
\description{
A dataset containing the relationships between ZIP code tabulation areas (ZCTA) and Census Tracts. This contains selected variables from the official relationship file. Built by \code{data-raw/03_build_zcta_crosswalk.R}; see \code{zip_data_version()} for the data release.
}
\keyword{datasets}
