% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_lookups.r
\name{zip_to_county}
\alias{zip_to_county}
\title{Returns the county name + state for a given ZIP code}
\usage{
zip_to_county(zip_code)
}
\arguments{
\item{zip_code}{A 5-digit U.S. ZIP code}
}
\value{
A named list consisting of the county name and state for the ZIP code
}
\description{
Returns the county name + state for a given ZIP code
}
\examples{
zip_to_county('90210')
zip_to_county('08731')
zip_to_county('07762')$county
#' zip_to_county('07762')$state
}
