% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.r
\name{download_zip_data}
\alias{download_zip_data}
\title{Download data files needed for library functionality}
\usage{
download_zip_data(force = FALSE)
}
\arguments{
\item{force}{Two-digit code for a U.S. state}
}
\value{
Data files needed for package functionality, stored in data directory of package install
}
\description{
Download data files needed for library functionality
}
\examples{

download_zip_data()


}
