% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_lookups.r
\name{geocode_zip}
\alias{geocode_zip}
\title{Returns that lat / lon pair of the centroid of a given ZIP code}
\usage{
geocode_zip(zip_code)
}
\arguments{
\item{zip_code}{A 5-digit U.S. ZIP code}
}
\value{
tibble of lat lon coordinates
}
\description{
Returns that lat / lon pair of the centroid of a given ZIP code
}
\examples{
geocode_zip("07762")
geocode_zip("90210")
geocode_zip("90210")$lat
geocode_zip("90210")$lng
}
