% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_lookups.r
\name{get_cd}
\alias{get_cd}
\title{Get all congressional districts for a given ZIP code}
\usage{
get_cd(zip_code)
}
\arguments{
\item{zip_code}{A U.S. ZIP code}
}
\value{
a named list of two-digit state code and two digit district code
}
\description{
Get all congressional districts for a given ZIP code
}
\examples{
get_cd('08731')
get_cd('90210')
}
