% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_lookups.r
\name{reverse_zipcode}
\alias{reverse_zipcode}
\title{Given a ZIP code, returns columns of metadata about that ZIP code}
\usage{
reverse_zipcode(zip_code)
}
\arguments{
\item{zip_code}{A 5-digit U.S. ZIP code or chracter vector with multiple ZIP codes}
}
\value{
A tibble containing data for the ZIP code(s)
}
\description{
Given a ZIP code, returns columns of metadata about that ZIP code
}
\examples{
reverse_zipcode("90210")
reverse_zipcode("08731")
reverse_zipcode(c("08734", "08731"))
reverse_zipcode("07762")$county
reverse_zipcode("07762")$state
}
