% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_lookups.r
\name{search_city}
\alias{search_city}
\title{Search ZIP codes for a city within a state}
\usage{
search_city(city_name, state_abb)
}
\arguments{
\item{city_name}{Name of major city to search}

\item{state_abb}{Two-digit code for a U.S. state}
}
\value{
dataframe of all ZIP code data found for given city
}
\description{
Search ZIP codes for a city within a state
}
\examples{
search_city('Spring Lake','NJ')
search_city('Chappaqua','NY')
}
