% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_lookups.r
\name{search_tz}
\alias{search_tz}
\title{Search ZIP codes for a timezone}
\usage{
search_tz(tz)
}
\arguments{
\item{tz}{Timezone}
}
\value{
dataframe of all ZIP codes found for given timezone
}
\description{
Search ZIP codes for a timezone
}
\examples{
eastern <- search_tz('Eastern')
pacific <- search_tz('Mountain')
}
