% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{zcta_crosswalk}
\alias{zcta_crosswalk}
\title{ZCTA to Census Tract (2010) Crosswalk}
\format{
A data frame with 148897 rows and 25 variables:
\describe{
\item{ZCTA5}{2010 ZIP Code Tabulation Area}
\item{STATE}{2010 State FIPS Code}
\item{COUNTY}{2010 County FIPS Code}
\item{TRACT}{2010 Census Tract Code}
\item{GEOID}{Concatenation of 2010 State, County, and Tract}
\item{POPPT}{Calculated 2010 Population for the relationship record}
\item{HUPT}{Calculated 2010 Housing Unit Count for the relationship record}
\item{AREAPT}{Total Area for the record}
\item{AREALANDPT}{Land Area for the record}
\item{ZPOP}{2010 Population of the 2010 ZCTA}
\item{ZHU}{2010 Housing Unit Count of the 2010 ZCTA}
\item{ZAREA}{Total Area of the 2010 ZCTA}
\item{ZAREALAND}{Total Land Area of the 2010 ZCTA}
\item{TRPOP}{2010 Population of the 2010 Census Tract}
\item{TRHU}{2010 Housing Unit Count of the 2010 Census Tract}
\item{TRAREA}{Total Area of the 2010 Census Tract}
\item{TRAREALAND}{Total Land Area of the 2010 Census Tract}
\item{ZPOPPCT}{The Percentage of Total Population of the 2010 ZCTA represented by the record}
\item{ZHUPCT}{The Percentage of Total Housing Unit Count of the 2010 ZCTA represented by the record}
\item{ZAREAPCT}{The Percentage of Total Area of the 2010 ZCTA represented by the record}
\item{ZAREALANDPCT}{The Percentage of Total Land Area of the 2010 ZCTA represented by the record}
\item{TRPOPPCT}{The Percentage of Total Population of the 2010 Census Tract represented by the record}
\item{TRHUPCT}{The Percentage of Total Housing Unit Count of the 2010 Census Tract represented by the record}
\item{TRAREAPCT}{The Percentage of Total Area of the 2010 Census Tract represented by the record}
\item{TRAREALANDPCT}{The Percentage of Total Land Area of the 2010 Census Tract represented by the record}
}
}
\source{
\url{https://www.census.gov/geographies/reference-files/time-series/geo/relationship-files.html}
}
\usage{
zcta_crosswalk
}
\description{
A dataset containing the relationships between ZIP code tabluation areas (ZCTA) and Census Tracts
}
\keyword{datasets}
