% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{zcta_crosswalk}
\alias{zcta_crosswalk}
\title{ZCTA to Census Tract (2010) Crosswalk}
\format{
A data frame with 148897 rows and 4 variables:
\describe{
\item{ZCTA5}{2010 ZIP Code Tabulation Area}
\item{TRACT}{2010 Census Tract Code}
\item{GEOID}{Concatenation of 2010 State, County, and Tract}
}
}
\source{
\url{https://www.census.gov/geographies/reference-files/time-series/geo/relationship-files.html}
}
\usage{
zcta_crosswalk
}
\description{
A dataset containing the relationships between ZIP code tabulation areas (ZCTA) and Census Tracts. This contains selected variables from the official crosswalk file.
}
\keyword{datasets}
