% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{zip_to_cd}
\alias{zip_to_cd}
\title{ZIP Code to Congressional District Relationship File}
\format{
A data frame with 45914 rows and 2 variables:
\describe{
\item{ZIP}{5 digit U.S. ZIP code}
\item{CD}{Four digit congressional district code (State FIPS code + district number)}
}
}
\source{
\url{https://www.huduser.gov/portal/datasets/usps_crosswalk.html}
}
\usage{
zip_to_cd
}
\description{
A dataset containing mappings between ZIP codes and congressional districts
}
\keyword{datasets}
